hardwere_configが命名上不適切だと感じたので修正
