--- Raspi/config/hardconfig.hpp
@@ -1,31 +1,0 @@
-//config/hardconfig.hpp
-//hardwareの基本設定に関するヘッダー
-#ifndef  HARDCONFIG_HPP
-    #define HARDCONFIG_HPP
-
-    #include<atomic>
-
-    class hardwere_config{
-        public:
-            class controller{
-                private:
-                    static inline std::atomic<bool> system_is_running{true};
-                
-                public:
-                    static std::atomic<bool>& get_system_is_running();
-                    /*
-                        system_is_running変数を安全に取得する目的
-                            引数：       void               
-                            返り値 ：   std::atomic<bool>&  system_is_runningのパス
-                    */
-                    static void put_system_is_running(bool system_flag);
-                    /*
-                        system_is_runningの値をsystem_flagの値で上書きする
-                        第1引数：   bool    system_flag      上書きする値
-                        返り値：    void   
-        
-                    */
-
-            };
-    };
-#endif
