+++ Raspi/src/app_tasks/clock_system.cpp
@@ -4,7 +4,7 @@
 
 
 #include"clock_system.hpp"
-#include"hardconfig.hpp"
+#include"softconfig.hpp"
 
 void clock_system::set_now_time(){
     while(true){
@@ -12,7 +12,7 @@ void clock_system::set_now_time(){
         //std::chronoによって時間を把握し、それをtime_t型に変換
         
 
-        if(hardwere_config::controller::get_system_is_running() == false)   break;
+        if(softwere_config::controller::get_system_is_running() == false)   break;
     }
     return;    
 }
